LPS-57110 Check for null

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression-evaluator/src/com/liferay/dynamic/data/mapping/expression/evaluator/internal/DDMExpressionEvaluatorConfigurationComponent.java
@@ -67,8 +67,13 @@ public class DDMExpressionEvaluatorConfigurationComponent {
 
 	@Deactivate
 	public void deactivate() throws IOException {
-		_restConfiguration.delete();
-		_cxfConfiguration.delete();
+		if (_restConfiguration != null) {
+			_restConfiguration.delete();
+		}
+
+		if (_cxfConfiguration != null) {
+			_cxfConfiguration.delete();
+		}
 	}
 
 	@Reference
